do not allow deletion of user id 1

+++ app/Models/Access/User/Traits/Attribute/UserAttribute.php
@@ -172,7 +172,7 @@ trait UserAttribute
      */
     public function getDeleteButtonAttribute()
     {
-        if ($this->id != access()->id()) {
+        if ($this->id != access()->id() && $this->id != 1) {
             return '<a href="'.route('admin.access.user.destroy', $this).'"
                  data-method="delete"
                  data-trans-button-cancel="'.trans('buttons.general.cancel').'"
